### 2022-06-09 #541 @smnnlt
spiro: Manage Data from Cardiopulmonary Exercise Testing
Submitting Author Name: Simon Nolte
Submitting Author Github Handle: @smnnlt
Repository:  https://github.com/smnnlt/spiro
Version submitted: 0.0.5
Submission type: Standard

Version accepted: TBD

---

```
Package: spiro
Title: Manage Data from Cardiopulmonary Exercise Testing
Version: 0.0.5
    person(given = "Simon",
           family = "Nolte",
           email = "s.nolte@dshs-koeln.de",
Description: Import, process, summarize and visualize raw data from
    metabolic carts.
    xml2,
    readxl,
    cowplot,
    digest,
    signal
    ggborderline
```

The spiro package allows to read and process data from raw data files of different metabolic carts.

This package is primarily written for researchers in exercise science, who want to make their analysis of cardiopulmonary exercise testing more standardized, reproducible and faster. It may also be used in a commercial context (e.g., training diagnostics business).

The [whippr](https://github.com/fmmattioni/whippr) package has a different approach to the same problem. Compared to whippr, spiro has a more automated and simpler data workflow (basically one function for reading and processing data, and one function for summarizing or plotting). spiro has several relevant additional features, that whippr does not have: Automated detection and manual generation of exercise test protocols; data summary by load steps; adding and synchronizing external heart rate data; import of raw data file meta data; advanced data filtering methods (e.g., Butterworth filters; moving breath averages); Wasserman 9-Panel-Plots. Compared to whippr, the spiro package does not offer methods for VO2 kinetics analysis and automated outlier removal.

This package works with cardiopulmonary exercise data, which is per se sensitive health data. Meta data from the original raw data files is read and anonymized by default (with the exception of data on body mass, which is necessary to perform certain calculations of variables). The anonymization can optionally be deactivated by means of a function argument [spiro(anonymize = FALSE)], so that meta data is saved alongside the processed data. This may be helpful in some settings when there is no intent to share the data. Sharing of the resulting data in such situations could potentially reveal personal information, which is why this option is not activated by default.

#537
