+++ NEWS.md
@@ -1,3 +1,9 @@
+spiro (development version)
+===========================
+
+
+
+
 spiro 0.0.5 (2022-06-07)
 ===========================
 
