### 2022-06-09 cfb64f1
Add examples to documentation
