+++ R/knit_print.R
@@ -20,6 +20,12 @@
 #' @param digits An integer giving the number of decimals to be rounded to.
 #' @param ... Passing of additional arguments to \code{knit_print.default()}.
 #'
+#' @examples
+#' # Get example data
+#' s <- spiro(spiro_example("zan_gxt"))
+#'
+#' knitr::knit_print(s)
+#' @export
 #' @importFrom knitr knit_print
 #' @export
 
