+++ man/spiro_interpolate.Rd
@@ -1,32 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/spiro_interpolate.R
-\name{spiro_interpolate}
-\alias{spiro_interpolate}
-\title{Interpolate spiroergometric data}
-\usage{
-spiro_interpolate(data)
-}
-\arguments{
-\item{data}{A \code{data.frame} containing the imported raw data. It is
-highly recommended to contain breath-by-breath data to avoid applying
-multiple interpolations on the original data.}
-}
-\value{
-A \code{data.frame} with data interpolated to every second.
-}
-\description{
-\code{spiro_interpolate()} linearly interpolates data from exercise testing
-to seconds.
-}
-\details{
-As spiroergometric data is usually recorded breath-by-breath it comprises
-different time intervals. For further calculations, a interpolation to every
-second of the measurement is beneficial.
-
-If there a duplicates in the time values (mostly due to rounding or imprecise
-time measurements) both values will be slightly separated to keep all
-information for the interpolation.
-
-Note that the applied linear interpolation will have a minor effect of
-smoothing on the data.
-}
