+++ man/spiro_import_vyntus.Rd
@@ -1,19 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/spiro_import.R
-\name{spiro_import_vyntus}
-\alias{spiro_import_vyntus}
-\title{Import raw data from Vyntus spiroergometric devices}
-\usage{
-spiro_import_vyntus(file)
-}
-\arguments{
-\item{file}{A character string, giving the path of the data file.}
-}
-\value{
-A \code{data.frame} with data. The attribute \code{info} contains
-addition meta-data retrieved from the original file.
-}
-\description{
-\code{spiro_import_vyntus()} retrieves cardiopulmonary data from Vyntus
-metabolic cart files.
-}
