### 2022-06-09 0c7b2ed
Rename helper functions for protocol setting
