+++ man/get_anonid.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/spiro_import.R
-\name{get_id}
-\alias{get_id}
+\name{get_anonid}
+\alias{get_anonid}
 \title{Get the anonymization id from personal data}
 \usage{
-get_id(name, surname, birthday = NULL)
+get_anonid(name, surname, birthday = NULL)
 }
 \arguments{
 \item{name}{A character string, containing the participant's name as present
@@ -21,7 +21,7 @@ data, this is ignored.}
 A character string, containing the anonymized id.
 }
 \description{
-\code{get_id()} returns the anonymization id corresponding to given personal
+\code{get_anonid()} returns the anonymization id corresponding to given personal
 data.
 }
 \details{
@@ -30,14 +30,14 @@ file meta data. The data are saved to the "info" attribute of a spiro() call.
 The default anonymization ensures that no personal information is
 accidentally revealed, e.g. by sharing spiro outputs as .Rda files.
 
-While there is no way to directly deanonymize the data, get_id() allows you
-to recreate the ids, when meta data (name, surname and birthday) are known.
-Birthday is only used within the id generation if available in the original
-raw data.
+While there is no way to directly deanonymize the data, get_anonid() allows
+you to recreate the ids, when meta data (name, surname and birthday) are
+known. Birthday is only used within the id generation if available in the
+original raw data.
 
 To disable the anonymization process during import use spiro(anonymize =
 FALSE)
 }
 \examples{
-get_id("Jesse", "Owens", "12.09.1913")
+get_anonid("Jesse", "Owens", "12.09.1913")
 }
