+++ NAMESPACE
@@ -8,7 +8,7 @@ export(add_bodymass)
 export(add_hr)
 export(add_protocol)
 export(bw_filter)
-export(get_id)
+export(get_anonid)
 export(get_protocol)
 export(pt_const)
 export(pt_pre)
