+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   import_processing: import_processing.html
   summarizing_plotting: summarizing_plotting.html
-last_built: 2022-06-08T15:38Z
+last_built: 2022-06-09T12:30Z
 urls:
   reference: https://smnnlt.github.io/spiro/reference
   article: https://smnnlt.github.io/spiro/articles
