+++ pkgdown.yml
@@ -1,10 +1,10 @@
 pandoc: 2.7.3
-pkgdown: 2.0.3
+pkgdown: 2.0.4
 pkgdown_sha: ~
 articles:
   import_processing: import_processing.html
   summarizing_plotting: summarizing_plotting.html
-last_built: 2022-06-09T12:30Z
+last_built: 2022-06-17T11:05Z
 urls:
   reference: https://smnnlt.github.io/spiro/reference
   article: https://smnnlt.github.io/spiro/articles
