### 2022-07-10 #541 @jameshunterbr
[spiro.index.nb.zip](https://github.com/ropensci/software-review/files/9079021/spiro.index.nb.zip)
