### 2022-07-10 #541 @melvidoni
> [spiro.index.nb.zip](https://github.com/ropensci/software-review/files/9079021/spiro.index.nb.zip)

Hello @jameshunterbr. Please remember to use the corresponding form to submit your review.
