### 2022-07-11 #541 @jameshunterbr
Melina Vidoni
This message is eligible for Automatic Cleanup! ***@***.***) Add cleanup rule | More info spiro.index.nb.zip Hello @jameshunterbr. Please remember to use the corresponding form to submit your review.
Hi, Melina:
I'm confused. I used the package [pkgreviewr](https://github.com/ropensci-org/pkgreviewr) as recommended in the Guide for Reviewers to prepare the report. Is that incorrect. My report used the template included in that package. Certainly the output it generated was far in excess of what I would normally send as a review for an article for example. I compressed the result to a zip as that was the only format I had that the space for the review accepted. How should I have submitted the review?
Regards,
Jim

https://github.com/ropensci-org/pkgreviewr	https://github.com/ropensci-org/pkgreviewr	[GitHub - ropensci-org/pkgreviewr: R package to facilitate rOpenSci package reviews.
R package to facilitate rOpenSci package reviews. Contribute to ropensci-org/pkgreviewr development by creating an account on GitHub.](https://github.com/ropensci-org/pkgreviewr)	https://github.com/ropensci-org/pkgreviewr

[James Hunter](https://spikenow.com/r/a/?ref=spike-organic-signature&_ts=1mu1q0)	[1mu1q0]
____________________________________
Prof. James R. Hunter, PhDBioinformática
Laboratório de Retrovirologia
Escola Paulista de Medicina
Universidade Federal de São Paulo
Cel: (55-11) 9-5327-5656
Rua Pedro de Toledo 669
6º Andar Fundos
04039-032 São Paulo, SP, BRASIL
Fone (11) 5576-4834
___________________________________
