### 2022-07-19 #541 @manuramon
Hello! Here is my review.

-   [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work.

The `spiro` packages focus on cardiopulmonar data and aims to standardize data import - from different available formats - and data processing and visualization.

Just to note, there is already another package with the name [spiro](https://github.com/wjschne/spiro) developed for creating spirographs.

-   README file sets out the background and the objectives addressed by the package. The inclusion of which devices are supported by this package in the README file is well appreciated.

-   Installation procedures are covered. No information on other package dependencies are provided.

-   Vignettes run successfully locally. "Import & Processing" Vignette includes information on formats (devices) supported by this packages as well as other useful information on how to import this data. "Summarizing & Plotting" vignette explains how to get resume measures from raw data and how to plot it.

-   Documentation for all external functions are provided with examples. As a suggestion, some of the examples show the functionality of the functions without assigning the output to a target, which results in printing the entire (truncated) database on the screen. In my opinion, I think it would be preferable to display only the header of that output.

#### Package testing

-   Installation have worked well on Linux and macOS platforms (not tested on Windows, sorry).

-   Package testing with `devtools::check()` on `spiro` returned 0 errors.

-   Package testing with `pkgcheck` or `goodpractice::gp` result in some warnings on unit tests (81% of code lines are covered by test) and length of code lines in vignette files. The `pkgcheck` function has also retuned a R CMD check error in the `DESCRIPTION` file -\> Required fields missing or empty:'Author' 'Maintainer'

-   Functions included in the package works as expected. A suggestion for `spiro_plot` function would be to include and argument to define the layout of the plot. I think it is easy to implement and could help to obtain graphs in a desirable layout (3x3 by default), especially when a subset of the 9 graphs is plotted.

-   A doubt about the functionality of the package is whether the functions allow to vectorize the actions across participants. If I understood correctly, the example databases are data measured on the same individual (one only) and, for example, the correction for body mass is done by giving the body mass data of this individual. Is it possible to read databases containing information from several individuals and work with the functions per individual? Maybe this question does not make sense and in this field it is usual to work with each file separately.
