### 2022-07-19 #541 @melvidoni
@jameshunterbr You have to add a comment with the review completed as the templated, as the one we just lodged above.

You have to follow the [review template](https://devguide.ropensci.org/reviewtemplate.html#reviewtemplate) which is in point [7.1 of the reviewing guide](https://devguide.ropensci.org/reviewerguide.html). I can't open the file you uploaded.
