### 2022-07-26 185d41c
Truncate output in documentation examples
