+++ man/add_protocol.Rd
@@ -26,7 +26,8 @@ data from an exercise test.
 file <- spiro_example("zan_gxt")
 
 s <- spiro(file)
-add_protocol(s, set_protocol(pt_pre(60), pt_steps(300, 50, 50, 7, 30)))
+out <- add_protocol(s, set_protocol(pt_pre(60), pt_steps(300, 50, 50, 7, 30)))
+head(out)
 }
 \seealso{
 \link{set_protocol} for protocol setting with helper functions.
