+++ DESCRIPTION
@@ -15,7 +15,7 @@ BugReports: https://github.com/smnnlt/spiro/issues
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.2.0
+RoxygenNote: 7.2.1
 Imports: 
     ggplot2,
     xml2,
