### 2022-07-26 5358f33
Allow customizations of plot arrangement in spiro_plot()
