### 2022-07-29 #541 @smnnlt
Hi,

I've started working on changes in response to the comments in the [dev-branch](https://github.com/smnnlt/spiro/tree/dev) of the package. I'll wait for @jameshunterbr to submit the review in the desired form before I'll respond to all comments here.
