### 2022-08-05 #541 @manuramon
Hi @melvidoni and @smnnlt , I am pasting below the review submitted by @jameshunterbr in the zip file, in case this help you. If finally @jameshunterbr uploads the review in a proper format, I will remove this comment.

All the best

---
output:
  html_document:
    toc: yes
    df_print: paged
  html_notebook:
    toc: yes
    toc_float: yes
editor_options:
  chunk_output_type: inline
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(
  collapse = TRUE,
  comment = "#>",
  fig.path = "man/figures/README-",
  out.width = "100%"
)

library(magrittr)
library(devtools)
```

# `spiro` - package review

## **Reviewer:** [\@jameshunterbr](https://github.com/jameshunterbr)

### Review Submitted:

**`r cat(sprintf("**Last updated:** %s", Sys.Date()))`**

This report contains documents associated with the review of **rOpenSci** submitted package:

## Package info

**Description:**

Import, process, summarize and visualize raw data from metabolic carts.

**Author:** `r      person(given = "Simon",            family = "Nolte",            role = c("aut", "cre"),            email = "s.nolte@dshs-koeln.de",            comment = c(ORCID = "0000-0003-1643-1860"))`

**repo url:** <https://github.com/smnnlt/spiro>

**website url:** <https://smnnlt.github.io/spiro>

## Review info

#### See [reviewer guidelines](https://ropensci.github.io/dev_guide/reviewerguide.html) for further information on the rOpenSci review process.

**key review checks:**

-   Does the code comply with **general principles in the [Mozilla reviewing guide](https://mozillascience.github.io/codeReview/review.html)**?
-   Does the package **comply with the [ROpenSci packaging guide](https://ropensci.github.io/dev_guide/building.html)**?
-   Are there **improvements** that could be made to the **code style?**
-   Is there **code duplication** in the package that should be reduced?
-   Are there **user interface improvements** that could be made?
-   Are there **performance improvements** that could be made?
-   Is the [**documentation**](https://ropensci.github.io/dev_guide/building.html#documentation) (installation instructions/vignettes/examples/demos) **clear and sufficient**?

Please be respectful and kind to the authors in your reviews. The rOpenSci [code of conduct](https://ropensci.github.io/dev_guide/policies.html#code-of-conduct) is mandatory for everyone involved in our review process.

### session info

```{r sessionInfo}
sessionInfo()
```

```{r pkg_dir, echo = F}
pkg_dir <- "C:/Users/james/OneDrive/Documents/R/rOpenSci/spiro"
```

## Test installation

### test local `spiro` install:

```{r test-local}
install(pkg_dir, dependencies = T, build_vignettes = T)
```

```{r github-rm}
remove.packages("spiro")
```

#### **comments:**

Local install worked correctly.

```{r test-github}
devtools::install_github("smnnlt/spiro", dependencies = T, build_vignettes = T)
```

#### **comments:**

Worked correctly.

## Check package integrity

### run checks on `spiro` source:

```{r check-checks}
devtools::check(pkg_dir)
```

#### **comments:**

OK

### run tests on `spiro` source:

```{r check-tests}
devtools::test(pkg_dir)
```

#### **comments:**

These tests encountered some problems and terminated early.

### check `spiro` for goodpractice:

```{r test-goodpractice}
goodpractice::gp(pkg_dir)
```

#### **comments:**

There are some recommendations here to be followed.

## Check package metadata files

### inspect

-

    #### [README](https://github.com/smnnlt/spiro)

-

    #### [DESCRIPTION](https://github.com/smnnlt/spiro/blob/master/DESCRIPTION)

-

### spell check

```{r spell-check}
devtools::spell_check(pkg_dir)
```

#### **comments:**

I don't know if you want to use the package outside the context of the NHS, but attention to the potential of use by professionals in other countries and language areas would be a good step to enhance the utility of your package.

## Check documentation

online documentation: [**https://smnnlt.github.io/spiro**](https://smnnlt.github.io/spiro){.uri}

-   Is the [documentation](https://ropensci.github.io/dev_guide/building.html#documentation) (installation instructions/vignettes/examples/demos) clear and sufficient?
    -   I had no problems installing the package and its subsidiary files from the instructions on GitHub.

### test `spiro` function help files:

```{r test-help}
detach(package:spiro,unload=TRUE)  # change versions of package
remotes::install_github("smnnlt/spiro@v0.0.5")
help(package = "spiro")
```

#### **comments:**

In the newest version, the response to the above command I received is the following:

`URL '/help/library/spiro/html/00Index.html' not found`

However, if I return to the 0.0.5 "stable version", the package help screen appears

As well, the request for help on the spiro() command failed in the new version (xx.9000):

`URL '/help/library/spiro/help/spiro' not found`

Likewise, the 0.0.5 version produces the help screen for the spiro() command

### test `spiro` vignettes:

```{r test-vignettes}
vignette(package = "spiro")
```

#### **comments:**

A package without vignettes is pretty naked. Some work here would pay big benefits.

**NOTE**: At this point I am going to stay with the 0.0.5 version instead of today's 0.0.5.9000 version as it obviously from the previous tests is more ready for use.

## Test functionality:

-   Are there **user interface improvements** that could be made?
-   Are there **performance improvements** that could be made?

```{r free-style}
library("spiro")
```

```{r parse-functions}
exports <-ls("package:spiro")
exports
```

```{r exp-chunk}

x <- spiro(spiro_example("zan_gxt"))
x
spiro_max(x)
spiro_plot(x)
imp <- spiro_import(spiro_example("zan_gxt"))
get_protocol(imp)
guess_device(imp)
```

#### **comments:**

2 points:

1.  a number of commands on the master help page (not the list above) are not yet implemented. E.g., "guess_device" although they have help pages. I would remove these from the help documentation until you operationalize them.
2.  You have some commands with very common command names like "pre" and "steps". I would rename these to something more specific to your project to avoid conflicts with other packages.

However, in general, your functions work as advertised in the help pages. You seem to have conquered to the hardest functions to implement.

## Inspect code:

-   Does the package **comply with the [ROpenSci packaging guide](https://ropensci.github.io/dev_guide/building.html)**?
    -   good [function & variable naming?](https://ropensci.github.io/dev_guide/building.html#function-and-argument-naming)
    -   good [dependency management](https://ropensci.github.io/dev_guide/building.html#package-dependencies)?
-   Are there **improvements** that could be made to the [**code style?**](https://ropensci.github.io/dev_guide/building.html#code-style)
-   Is there **code duplication** in the package that should be reduced?

```{r inspect-code}
pkgreviewr::pkgreview_print_source("spiro")
```

**\* might not be suitable for large packages with many exported functions**

<br> <br>

#### **comments:**

Here the unimplemented commands are clearly stated. I was able to follow almost all of the code and it appears to accomplish what you set out to do.

## Review test suite:

See guidance on [testing](https://ropensci.github.io/dev_guide/building.html#testing) for further details.

### test coverage

```{r pkg_coverage}
covr::package_coverage(pkg_dir)

```

### inspect [tests](https://github.com/smnnlt/spiro/blob/master/tests/testthat)

#### **comments:**

The tests appear to cover the principal functions of the package and are complete. However, I don't see why my system is not testing the coverage. Sorry.
