### 2022-08-07 #541 @jameshunterbr
## Package Review

*Please check off boxes as applicable, and elaborate in comments below. Your review is not limited to these topics, as described in the reviewer guide\*

- *Briefly describe any working relationship you have (had) with the package authors.*

- [X] As the reviewer I confirm that there are no [conflicts of interest](%5Bhttps://devguide.ropensci.org/policies.html#coi)](<https://devguide.ropensci.org/policies.html#coi>)) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

#### Documentation

- [X ] *A statement of need:* clearly stating problems the software is designed to solve and its target audience in README

- [X ] *Installation instructions:* for the development version of package and any non-standard dependencies in README

- [ ] *Vignette(s):* demonstrating major functionality that runs successfully locally

- [X ] *Function Documentation:* for all exported functions

- [X ] *Examples:* (that run successfully locally) for all exported functions

- [ ] *Community guidelines:* including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

#### Functionality

- [X ] *Installation:* Installation succeeds as documented.

- [ ] *Functionality:* Any functional claims of the software been confirmed.

- [ ] *Performance:* Any performance claims of the software been confirmed.

- [ ] *Automated tests:* Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.

- [ ] *Packaging guidelines*: The package conforms to the rOpenSci packaging guidelines.

Estimated hours spent reviewing:6

- [X ] Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

---

### Review Comments

1. My first observation is that all the functions listed in the main help screen have not yet been programmed. So either those that are unfinished should be removed and re-inserted in a later version when they are completed.

2. Likewise, there are no vignettes and there should be. In the README file you have examples, but they should be more complete so that someone in some outlying NHS locale can use them without constant referral to you.

3. In the newest version, the response to the installation command (remotes::install_github("smnnlt/spiro@v0.0.5") I received is the following:

`URL '/help/library/spiro/html/00Index.html' not found`

However, if I return to the 0.0.5 "stable version", the package help screen appears.

As well, the request for help on the spiro() command failed in the new version (xx.9000):

`URL '/help/library/spiro/help/spiro' not found`

Likewise, the 0.0.5 version produces the help screen for the spiro() command

4. You have some commands with very common command names like "pre" and "steps". I would rename these to something more specific to your project to avoid conflicts with other packages.

5. Regarding tests, they appear to cover the principal functions of the package and are complete. However, my system is not testing the coverage for some reason. Sorry.
