### 2022-08-07 #541 @jameshunterbr
Sorry about the delay in reformatting and resubmitting the review of
Spiro. I hope this meets your format.

However, the previous version I uploaded was in response to an R package
suggested in your Reviewer's Guidelines. I don't know how I was supposed
to know that that format was not correct.

Sincerely,

James Hunter

—————————————————
James R. Hunter, Ph.D.
Laboratório de Retrovirologia
Disciplina de Infectologia
Departamento de Medicina
Escola Paulista de Medicina
Universidade Federal de São Paulo
Cel: (11) 9-5327-5656

Rua Pedro de Toledo 669
6º Andar Fundos
04039-032  São Paulo, SP, BRASIL
Fone (11) 5576-4834
—————————————————
