### 2022-08-08 #541 @smnnlt
Hi @manuramon @jameshunterbr

Thank you very much for your reviews! I'll respond to all issues raised by each report:

### Response to @manuramon

> As a suggestion, some of the examples show the functionality of the functions without assigning the output to a target, which results in printing the entire (truncated) database on the screen. In my opinion, I think it would be preferable to display only the header of that output. An alternative solution could be to modify the `print.spiro` function to show only a set of rows.

Thanks for the comment. I have rewritten the examples to to print only truncated output using the head() function (smnnlt/spiro@185d41c).

> Package testing with `pkgcheck` or `goodpractice::gp` result in some warnings on unit tests (81% of code lines are covered by test) and length of code lines in vignette files. The `pkgcheck` function has also retuned a R CMD check error in the `DESCRIPTION` file -> Required fields missing or empty:'Author' 'Maintainer'

I shortened the code lines in the vignettes (smnnlt/spiro@32597b5). I do not know the source for the R CMD error, since `devtools::check()` and R CMD check via CI do not throw any errors.

> Functions included in the package works as expected. A suggestion for `spiro_plot` function would be to include and argument to define the layout of the plot. I think it is easy to implement and could help to obtain graphs in a desirable layout (3x3 by default), especially when a subset of the 9 graphs is plotted.

This is a great idea! I have added an argument `grid_args` to the `spiro_plot()` function, which passes arguments to the internal `cowplot::plot_grid()` call (smnnlt/spiro@5358f33). This allows to modify the layout of the plot arrangement (e.g. size, labels, alignment).

> A doubt about the functionality of the package is whether the functions allow to vectorize the actions across participants. If I understood correctly, the example databases are data measured on the same individual (one only) and, for example, the correction for body mass is done by giving the body mass data of this individual. Is it possible to read databases containing information from several individuals and work with the functions per individual? Maybe this question does not make sense and in this field it is usual to work with each file separately.

You're completely right, the example database is only data for a single individual and in some cases the user may want to work with data from several individuals at once. The spiro package does currently not support native vectorization, but such functionality is easy to achieve using packages for vectorization such as {purrr}. For example, the case you mentioned (simultaneously vectorizing over individuals' data files and body mass data) could be dealt with `purrr::map2(.x = list_of_rawdata, .y = list_of_bodymass, .f = spiro)`. I thought about adding such an example to the vignettes (or creating a new one), but I don't want the package to have a dependency on {purrr}. I may write this use case up in a blog post or embedded it into the article I plan to submit to JOSS.

### Response to @jameshunterbr

> My first observation is that all the functions listed in the main help screen have not yet been programmed. So either those that are unfinished should be removed and re-inserted in a later version when they are completed.

All exported functions should now have documentation pages including examples (smnnlt/spiro@cfb64f1).

> Likewise, there are no vignettes and there should be. In the README file you have examples, but they should be more complete so that someone in some outlying NHS locale can use them without constant referral to you.

The package has two vignettes (also available on its [website](https://smnnlt.github.io/spiro/)). To be able to view them locally, be sure to state `build_vignettes = TRUE` when running `remotes::install_github()`. I have added some more context to the README (smnnlt/spiro@68e739a). Further examples can be found in the vignettes.

> In the newest version, the response to the installation command `remotes::install_github("smnnlt/spiro@v0.0.5")` I received is the following: `URL '/help/library/spiro/html/00Index.html' not found`. However, if I return to the 0.0.5 "stable version", the package help screen appears. As well, the request for help on the `spiro()` command failed in the new version (xx.9000): `URL '/help/library/spiro/help/spiro' not found`.

I'm sorry, but I could not reproduce these issues. On my devices, the package successfully builds the documentation pages when being installed.

> You have some commands with very common command names like "pre" and "steps". I would rename these to something more specific to your project to avoid conflicts with other packages.

Thank for pointing out this important issue! I have renamed the functions that may cause name conflicts with popular packages (smnnlt/spiro@0c7b2ed & smnnlt/spiro@0199249).

> Regarding tests, they appear to cover the principal functions of the package and are complete. However, my system is not testing the coverage for some reason. Sorry.

That's no problem. You may visit the project's [Codecov](https://app.codecov.io/gh/smnnlt/spiro) page to get an impression on the package's code coverage.

Thanks again for both of your reports @manuramon @jameshunterbr . I have added you to the `DESCRIPTION` as package reviewers  (smnnlt/spiro@a980365). If you're fine with it, I would also add you to the 'Acknowledgement' section in the `README`.
