### 2022-08-09 #541 @jameshunterbr
I have looked through the new changes that Simon has made and he has responded to my comments well. I would think he might be ready to change versions. What I see now, especially in terms of documentation, is much better and ready for publication.
