### 2022-08-17 #541 @manuramon
Hi all.
Thanks @smnnlt for considering some of my suggestions. I have retested the `spiro` package and everything works without errors. I have also tested the `grid_args` argument in the `spiro_plot` function. As I told you before, in my opinion it adds more flexibility in the plotting functionality. The example added in the `spiro_plot` help helps to understand how this new function works.

@melvidoni I have nothing more to add to this review.

All the best.
