### 2022-08-17 #541 @ropensci-review-bot
Approved! Thanks @smnnlt for submitting and @jameshunterbr, @manuramon for your reviews! :grin:
