### 2022-10-11 #541 @jameshunterbr
Thanks for responding to the issues we raised and for the mention in the
DESCRIPTION file. It would be fine to add me as a reviewer in the
Acknowledgement section. Thank you.

Best regards,

James Hunter

—————————————————
James R. Hunter, Ph.D.
Laboratório de Retrovirologia
Disciplina de Infectologia
Departamento de Medicina
Escola Paulista de Medicina
Universidade Federal de São Paulo
Cel: (11) 9-5327-5656

Rua Pedro de Toledo 669
6º Andar Fundos
04039-032  São Paulo, SP, BRASIL
Fone (11) 5576-4834
—————————————————
